fix: extract Genre/Rating controllers from FilmController
