+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -1,0 +1,28 @@
+package ru.yandex.practicum.filmorate.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import ru.yandex.practicum.filmorate.model.enums.Genre;
+import ru.yandex.practicum.filmorate.service.FilmService;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/genres")
+@RequiredArgsConstructor
+public class GenreController {
+    private final FilmService filmService;
+
+    @GetMapping
+    public List<Genre> getGenres() {
+        return filmService.getGenres();
+    }
+
+    @GetMapping("/{genreId}")
+    public Genre getGenreById(@PathVariable Integer genreId) {
+        return filmService.getGenreById(genreId);
+    }
+}
