+++ src/main/resources/application.yaml
@@ -3,7 +3,8 @@ spring:
     init:
       mode: always
   datasource:
-    url: jdbc:h2:file:./db/filmorate
+    url: jdbc:h2:mem:testDb
+#    jdbc:h2:file:./db/filmorate
     driverClassName: org.h2.Driver
     username: sa
     password: password
